Refactor analysis helpers to new module
